Refactor createConsumer to Consumer
