+++ packages/counter-example/src/App.js
@@ -1,6 +1,7 @@
 import React from 'react'
 
-import Context from './context'
+import { Provider } from './context'
+
 import Up from './context/Up'
 import Down from './context/Down'
 import Count from './context/Count'
@@ -8,7 +9,7 @@ import TotalCount from './context/TotalCount'
 import DownOnly from './context/DownOnly'
 
 const App = () => (
-  <Context.Provider>
+  <Provider>
     <Up />
 
     <Down />
@@ -18,7 +19,7 @@ const App = () => (
     <TotalCount />
 
     <DownOnly />
-  </Context.Provider>
+  </Provider>
 )
 
 export default App
