+++ packages/counter-example/src/context/Count.js
@@ -1,10 +1,11 @@
 import React from 'react'
 
-import Context from './index'
+import { Consumer } from './index'
 
 const mapState = ({ counter }) => counter
 
-const Consumer = Context.createConsumer(mapState)
-const Count = () => <Consumer>{state => <h3>Count: {state}</h3>}</Consumer>
+const Count = () => (
+  <Consumer mapState={mapState}>{state => <h3>Count: {state}</h3>}</Consumer>
+)
 
 export default Count
