+++ packages/counter-example/src/context/Down.js
@@ -1,12 +1,13 @@
 import React from 'react'
 
-import Context from './index'
+import { Consumer } from './index'
 
 const mapActions = ({ countDown }) => countDown
 
-const Consumer = Context.createConsumer(false, mapActions)
 const Down = () => (
-  <Consumer>{(_, action) => <button onClick={action}>DOWN</button>}</Consumer>
+  <Consumer mapActions={mapActions}>
+    {(_, action) => <button onClick={action}>DOWN</button>}
+  </Consumer>
 )
 
 export default Down
