+++ packages/counter-example/src/context/DownOnly.js
@@ -1,10 +1,11 @@
 import React from 'react'
 
-import Context from './index'
+import { Consumer } from './index'
 
 const mapState = ({ downs }) => downs
 
-const Consumer = Context.createConsumer(mapState)
-const DownsOnly = () => <Consumer>{state => <h3>Downs: {state}</h3>}</Consumer>
+const DownsOnly = () => (
+  <Consumer mapState={mapState}>{state => <h3>Downs: {state}</h3>}</Consumer>
+)
 
 export default DownsOnly
