+++ packages/counter-example/src/context/TotalCount.js
@@ -1,12 +1,13 @@
 import React from 'react'
 
-import Context from './index'
+import { Consumer } from './index'
 
 const mapState = ({ total }) => total
 
-const Consumer = Context.createConsumer(mapState)
 const TotalCount = () => (
-  <Consumer>{state => <h3>Total count: {state}</h3>}</Consumer>
+  <Consumer mapState={mapState}>
+    {state => <h3>Total count: {state}</h3>}
+  </Consumer>
 )
 
 export default TotalCount
